+++ src/Cli/func/Diagnostics/LoggingFilterHelper.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
+using System;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 
@@ -15,7 +16,7 @@ namespace Azure.Functions.Cli
         public const string CiBuildNumber = "BUILD_NUMBER";  // Travis CI, Cirrus CI
         public const string CiRunId = "RUN_ID"; // TaskCluster, dsari
 
-        public LoggingFilterHelper(IConfigurationRoot hostJsonConfig, bool? verboseLogging)
+        public LoggingFilterHelper(IConfigurationRoot hostJsonConfig, bool? verboseLogging, string userLogLevel = null)
         {
             VerboseLogging = verboseLogging.HasValue && verboseLogging.Value;
 
@@ -29,10 +30,31 @@ namespace Azure.Functions.Cli
                 SystemLogDefaultLogLevel = LogLevel.Information;
             }
 
+            // Check for environment variable if userLogLevel is not provided via CLI
+            if (string.IsNullOrEmpty(userLogLevel))
+            {
+                userLogLevel = Environment.GetEnvironmentVariable("FUNCTIONS_USER_LOG_LEVEL");
+            }
+
+            // Track if userLogLevel was explicitly set and successfully parsed
+            bool userLogLevelExplicitlySet = false;
+
+            // If userLogLevel is specified (via CLI or env var), use it
+            if (!string.IsNullOrEmpty(userLogLevel) && Enum.TryParse<LogLevel>(userLogLevel, true, out LogLevel parsedUserLogLevel))
+            {
+                UserLogDefaultLogLevel = parsedUserLogLevel;
+                userLogLevelExplicitlySet = true;
+            }
+
             if (Utilities.LogLevelExists(hostJsonConfig, Utilities.LogLevelDefaultSection, out LogLevel logLevel))
             {
                 SystemLogDefaultLogLevel = logLevel;
-                UserLogDefaultLogLevel = logLevel;
+
+                // Only override UserLogDefaultLogLevel if it wasn't explicitly set via CLI or env var
+                if (!userLogLevelExplicitlySet)
+                {
+                    UserLogDefaultLogLevel = logLevel;
+                }
             }
         }
 
@@ -49,7 +71,7 @@ namespace Azure.Functions.Cli
         /// <summary>
         /// Gets a value indicating whether is set to true if `func start` is started with `--verbose` flag. If set, SystemLogDefaultLogLevel is set to Information.
         /// </summary>
-        public bool VerboseLogging { get; private set; }
+        public bool VerboseLogging { get; }
 
         internal bool IsCiEnvironment(bool verboseLoggingArgExists)
         {